requirements.txt
python-telegram-bot==20.7
python-telegram-bot[webhooks]==20.7
sqlalchemy==2.0.25
aiosqlite==0.19.0
python-dotenv==1.0.0
